+++ engine/src/EnsembleEngine.h
@@ -36,64 +36,13 @@
 #include <vector>
 #include <assert.h>
 
+#include "MetaEngine.h"
 #include "BooleanNetwork.h"
 #include "Cumulator.h"
 #include "RandomGenerator.h"
 #include "RunConfig.h"
 
 struct EnsembleArgWrapper;
-
-class MetaEngine {
-
-protected:
-
-  RunConfig* runconfig;
-
-  double time_tick;
-  double max_time;
-  unsigned int sample_count;
-  bool discrete_time;
-  unsigned int thread_count;
-  
-  NetworkState reference_state;
-  unsigned int refnode_count;
-
-  mutable long long elapsed_core_runtime, user_core_runtime, elapsed_statdist_runtime, user_statdist_runtime, elapsed_epilogue_runtime, user_epilogue_runtime;
-  STATE_MAP<NetworkState_Impl, unsigned int> fixpoints;
-  std::vector<STATE_MAP<NetworkState_Impl, unsigned int>*> fixpoint_map_v;
-  
-  Cumulator* merged_cumulator;
-  std::vector<Cumulator*> cumulator_v;
-
-  STATE_MAP<NetworkState_Impl, unsigned int>* mergeFixpointMaps();
-
-public:
-
-  MetaEngine(RunConfig* runconfig) : 
-    runconfig(runconfig),
-    time_tick(runconfig->getTimeTick()), 
-    max_time(runconfig->getMaxTime()), 
-    sample_count(runconfig->getSampleCount()), 
-    discrete_time(runconfig->isDiscreteTime()), 
-    thread_count(runconfig->getThreadCount()) {}
-
-  static void init();
-  static void loadUserFuncs(const char* module);
-
-  long long getElapsedCoreRunTime() const {return elapsed_core_runtime;}
-  long long getUserCoreRunTime() const {return user_core_runtime;}
-
-  long long getElapsedEpilogueRunTime() const {return elapsed_epilogue_runtime;}
-  long long getUserEpilogueRunTime() const {return user_epilogue_runtime;}
-
-  long long getElapsedStatDistRunTime() const {return elapsed_statdist_runtime;}
-  long long getUserStatDistRunTime() const {return user_statdist_runtime;}
-
-  bool converges() const {return fixpoints.size() > 0;}
-  const STATE_MAP<NetworkState_Impl, unsigned int>& getFixpoints() const {return fixpoints;}
-
-};
-
 
 class EnsembleEngine : MetaEngine {
