+++ engine/src/MaBEstEngine.cc
@@ -40,36 +40,10 @@
 
 const std::string MaBEstEngine::VERSION = "2.3.0";
 size_t RandomGenerator::generated_number_count = 0;
-static const char* MABOSS_USER_FUNC_INIT = "maboss_user_func_init";
-
-void MaBEstEngine::init()
-{
-  extern void builtin_functions_init();
-  builtin_functions_init();
-}
-
-void MaBEstEngine::loadUserFuncs(const char* module)
-{
-  init();
-
-  void* dl = dlopen(module, RTLD_LAZY);
-  if (NULL == dl) {
-    std::cerr << dlerror() << "\n";
-    exit(1);
-  }
-
-  void* sym = dlsym(dl, MABOSS_USER_FUNC_INIT);
-  if (sym == NULL) {
-    std::cerr << "symbol " << MABOSS_USER_FUNC_INIT << "() not found in user func module: " << module << "\n";
-    exit(1);
-  }
-  typedef void (*init_t)(std::map<std::string, Function*>*);
-  init_t init_fun = (init_t)sym;
-  init_fun(Function::getFuncMap());
-}
 
 MaBEstEngine::MaBEstEngine(Network* network, RunConfig* runconfig) :
-  network(network), runconfig(runconfig), time_tick(runconfig->getTimeTick()), max_time(runconfig->getMaxTime()), sample_count(runconfig->getSampleCount()), discrete_time(runconfig->isDiscreteTime()), thread_count(runconfig->getThreadCount()) {
+  MetaEngine(network, runconfig)
+  {
 
   if (thread_count > sample_count) {
     thread_count = sample_count;
@@ -423,229 +397,9 @@ void MaBEstEngine::display(std::ostream& output_probtraj, std::ostream& output_s
   }
 }
 
-const std::map<unsigned int, std::pair<NetworkState, double> > MaBEstEngine::getFixPointsDists() const {
-  
-  std::map<unsigned int, std::pair<NetworkState, double> > res;
-  if (0 == fixpoints.size()) {
-    return res;
-  }
-
-  STATE_MAP<NetworkState_Impl, unsigned int>::const_iterator begin = fixpoints.begin();
-  STATE_MAP<NetworkState_Impl, unsigned int>::const_iterator end = fixpoints.end();
-  
-  for (unsigned int nn = 0; begin != end; ++nn) {
-    const NetworkState& network_state = (*begin).first;
-    res[nn] = std::make_pair(network_state,(double) (*begin).second / sample_count);
-    ++begin;
-  }
-  return res;
-}
-
-
 void MaBEstEngine::displayAsymptotic(std::ostream& output_asymptprob, bool hexfloat, bool proba) const
 {
   merged_cumulator->displayAsymptoticCSV(network, refnode_count, output_asymptprob, hexfloat, proba);
-}
-
-const std::map<double, STATE_MAP<NetworkState_Impl, double> > MaBEstEngine::getStateDists() const {
-  return merged_cumulator->getStateDists();
-}
-
-const STATE_MAP<NetworkState_Impl, double> MaBEstEngine::getNthStateDist(int nn) const {
-  return merged_cumulator->getNthStateDist(nn);
-}
-
-const STATE_MAP<NetworkState_Impl, double> MaBEstEngine::getAsymptoticStateDist() const {
-  return merged_cumulator->getAsymptoticStateDist();
-}
-
-const std::map<double, std::map<Node *, double> > MaBEstEngine::getNodesDists() const {
-
-  std::map<double, std::map<Node *, double> > result;
-
-  const std::map<double, STATE_MAP<NetworkState_Impl, double> > state_dist = merged_cumulator->getStateDists();
-
-  std::map<double, STATE_MAP<NetworkState_Impl, double> >::const_iterator begin = state_dist.begin();
-  std::map<double, STATE_MAP<NetworkState_Impl, double> >::const_iterator end = state_dist.end();
-  
-  const std::vector<Node*>& nodes = network->getNodes();
-
-  while(begin != end) {
-
-    std::map<Node *, double> node_dist;
-    STATE_MAP<NetworkState_Impl, double> present_state_dist = begin->second;
-
-    std::vector<Node *>::const_iterator nodes_begin = nodes.begin();
-    std::vector<Node *>::const_iterator nodes_end = nodes.end();
-
-    while(nodes_begin != nodes_end) {
-
-      if (!(*nodes_begin)->isInternal())
-      {
-        double dist = 0;
-
-        STATE_MAP<NetworkState_Impl, double>::const_iterator states_begin = present_state_dist.begin();
-        STATE_MAP<NetworkState_Impl, double>::const_iterator states_end = present_state_dist.end();
-      
-        while(states_begin != states_end) {
-
-          NetworkState state = states_begin->first;
-          dist += states_begin->second * ((double) state.getNodeState(*nodes_begin));
-
-          states_begin++;
-        }
-
-        node_dist[*nodes_begin] = dist;
-      }
-      nodes_begin++;
-    }
-
-    result[begin->first] = node_dist;
-
-    begin++;
-  }
-
-  return result;
-}
-
-const std::map<Node*, double> MaBEstEngine::getNthNodesDist(int nn) const {
-  std::map<Node *, double> result;
-
-  const STATE_MAP<NetworkState_Impl, double> state_dist = merged_cumulator->getNthStateDist(nn);
-  
-  const std::vector<Node*>& nodes = network->getNodes();
-  std::vector<Node *>::const_iterator nodes_begin = nodes.begin();
-  std::vector<Node *>::const_iterator nodes_end = nodes.end();
-
-  while(nodes_begin != nodes_end) {
-
-    if (!(*nodes_begin)->isInternal())
-    {
-      double dist = 0;
-
-      STATE_MAP<NetworkState_Impl, double>::const_iterator states_begin = state_dist.begin();
-      STATE_MAP<NetworkState_Impl, double>::const_iterator states_end = state_dist.end();
-    
-      while(states_begin != states_end) {
-
-        NetworkState state = states_begin->first;
-        dist += states_begin->second * ((double) state.getNodeState(*nodes_begin));
-
-        states_begin++;
-      }
-
-      result[*nodes_begin] = dist;
-    }
-    nodes_begin++;
-  }
-
-  return result;  
-}
-
-const std::map<Node*, double> MaBEstEngine::getAsymptoticNodesDist() const {
-  std::map<Node *, double> result;
-
-  const STATE_MAP<NetworkState_Impl, double> state_dist = merged_cumulator->getAsymptoticStateDist();
-  
-  const std::vector<Node*>& nodes = network->getNodes();
-  std::vector<Node *>::const_iterator nodes_begin = nodes.begin();
-  std::vector<Node *>::const_iterator nodes_end = nodes.end();
-
-  while(nodes_begin != nodes_end) {
-
-    if (!(*nodes_begin)->isInternal())
-    {
-      double dist = 0;
-
-      STATE_MAP<NetworkState_Impl, double>::const_iterator states_begin = state_dist.begin();
-      STATE_MAP<NetworkState_Impl, double>::const_iterator states_end = state_dist.end();
-    
-      while(states_begin != states_end) {
-
-        NetworkState state = states_begin->first;
-        dist += states_begin->second * ((double) state.getNodeState(*nodes_begin));
-
-        states_begin++;
-      }
-
-      result[*nodes_begin] = dist;
-    }
-    nodes_begin++;
-  }
-
-  return result;  
-}
-
-const std::map<double, double> MaBEstEngine::getNodeDists(Node * node) const {
- 
-  std::map<double, double> result;
-
-  const std::map<double, STATE_MAP<NetworkState_Impl, double> > state_dist = merged_cumulator->getStateDists();
-
-  std::map<double, STATE_MAP<NetworkState_Impl, double> >::const_iterator begin = state_dist.begin();
-  std::map<double, STATE_MAP<NetworkState_Impl, double> >::const_iterator end = state_dist.end();
-
-  while(begin != end) {
-
-    STATE_MAP<NetworkState_Impl, double> present_state_dist = begin->second;
-    double dist = 0;
-
-    STATE_MAP<NetworkState_Impl, double>::const_iterator states_begin = present_state_dist.begin();
-    STATE_MAP<NetworkState_Impl, double>::const_iterator states_end = present_state_dist.end();
-  
-    while(states_begin != states_end) {
-
-      NetworkState state = states_begin->first;
-      dist += states_begin->second * ((double) state.getNodeState(node));
-
-      states_begin++;
-    }
-    result[begin->first] = dist;
-
-    begin++;
-  }
-
-  return result; 
-}
-
-double MaBEstEngine::getNthNodeDist(Node * node, int nn) const {
-
-  double result = 0;
-
-  const STATE_MAP<NetworkState_Impl, double> state_dist = merged_cumulator->getNthStateDist(nn);
-  
-  STATE_MAP<NetworkState_Impl, double>::const_iterator states_begin = state_dist.begin();
-  STATE_MAP<NetworkState_Impl, double>::const_iterator states_end = state_dist.end();
-
-  while(states_begin != states_end) {
-
-    NetworkState state = states_begin->first;
-    result += states_begin->second * ((double) state.getNodeState(node));
-
-    states_begin++;
-  }
-
-  return result;  
-}
-
-double MaBEstEngine::getAsymptoticNodeDist(Node * node) const {
-
-  double result = 0;
-
-  const STATE_MAP<NetworkState_Impl, double> state_dist = merged_cumulator->getAsymptoticStateDist();
-  
-  STATE_MAP<NetworkState_Impl, double>::const_iterator states_begin = state_dist.begin();
-  STATE_MAP<NetworkState_Impl, double>::const_iterator states_end = state_dist.end();
-
-  while(states_begin != states_end) {
-
-    NetworkState state = states_begin->first;
-    result += states_begin->second * ((double) state.getNodeState(node));
-
-    states_begin++;
-  }
-
-  return result;  
 }
 
 MaBEstEngine::~MaBEstEngine()
