+++ engine/src/MaBEstEngine.h
@@ -37,6 +37,7 @@
 #include <vector>
 #include <assert.h>
 
+#include "MetaEngine.h"
 #include "BooleanNetwork.h"
 #include "Cumulator.h"
 #include "RandomGenerator.h"
@@ -44,23 +45,8 @@
 
 struct ArgWrapper;
 
-class MaBEstEngine {
+class MaBEstEngine : public MetaEngine {
 
-  Network* network;
-  RunConfig* runconfig;
-  double time_tick;
-  double max_time;
-  unsigned int sample_count;
-  bool discrete_time;
-  unsigned int thread_count;
-  //std::map<NetworkState, unsigned int> fixpoints;
-  STATE_MAP<NetworkState_Impl, unsigned int> fixpoints;
-  mutable long long elapsed_core_runtime, user_core_runtime, elapsed_statdist_runtime, user_statdist_runtime, elapsed_epilogue_runtime, user_epilogue_runtime;
-  NetworkState reference_state;
-  unsigned int refnode_count;
-  Cumulator* merged_cumulator;
-  std::vector<STATE_MAP<NetworkState_Impl, unsigned int>*> fixpoint_map_v;
-  std::vector<Cumulator*> cumulator_v;
   std::vector<ArgWrapper*> arg_wrapper_v;
   NodeIndex getTargetNode(RandomGenerator* random_generator, const MAP<NodeIndex, double>& nodeTransitionRates, double total_rate) const;
   double computeTH(const MAP<NodeIndex, double>& nodeTransitionRates, double total_rate) const;
@@ -71,46 +57,13 @@ class MaBEstEngine {
 
 public:
   static const std::string VERSION;
-  static void init();
-  static void loadUserFuncs(const char* module);
   
   MaBEstEngine(Network* network, RunConfig* runconfig);
 
   void run(std::ostream* output_traj);
 
-  //const std::map<NetworkState, unsigned int>& getFixpoints() const {return fixpoints;}
-  const STATE_MAP<NetworkState_Impl, unsigned int>& getFixpoints() const {return fixpoints;}
-
-  bool converges() const {return fixpoints.size() > 0;}
-
   void display(std::ostream& output_probtraj, std::ostream& output_statdist, std::ostream& output_fp, bool hexfloat = false) const;
   void displayAsymptotic(std::ostream& output_asymptprob, bool hexfloat = false, bool proba = true) const;
-
-  const std::map<double, STATE_MAP<NetworkState_Impl, double> > getStateDists() const;
-  const STATE_MAP<NetworkState_Impl, double> getNthStateDist(int nn) const;
-  const STATE_MAP<NetworkState_Impl, double> getAsymptoticStateDist() const;
-
-  const std::map<double, std::map<Node *, double> > getNodesDists() const;
-  const std::map<Node*, double> getNthNodesDist(int nn) const;
-  const std::map<Node*, double> getAsymptoticNodesDist() const;
-
-  const std::map<double, double> getNodeDists(Node * node) const;
-  double getNthNodeDist(Node * node, int nn) const;
-  double getAsymptoticNodeDist(Node * node) const;
-  
-  const std::map<unsigned int, std::pair<NetworkState, double> > getFixPointsDists() const;
-
-
-  int getMaxTickIndex() const {return merged_cumulator->getMaxTickIndex();} 
-
-  long long getElapsedCoreRunTime() const {return elapsed_core_runtime;}
-  long long getUserCoreRunTime() const {return user_core_runtime;}
-
-  long long getElapsedEpilogueRunTime() const {return elapsed_epilogue_runtime;}
-  long long getUserEpilogueRunTime() const {return user_epilogue_runtime;}
-
-  long long getElapsedStatDistRunTime() const {return elapsed_statdist_runtime;}
-  long long getUserStatDistRunTime() const {return user_statdist_runtime;}
 
   ~MaBEstEngine();
 };
